requirements.txt
requests>=2.22.0
websocket-client>=1.2.0